# docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Cart and Stock API
  version: 1.0.0
  description: Contrato HTTP das APIs de catálogo, estoque e carrinho.
servers:
  - url: /
security:
  - ApiKeyAuth: []
...
paths:
  /products:
    get:
      tags: [Catálogo]
      summary: Lista produtos do catálogo
      operationId: listProducts
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Produtos encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Filtro inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Catálogo]
      summary: Cria um produto com estoque zero
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Nome, slug ou preço inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Slug já utilizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /stocks/{productId}:
    get:
      tags: [Estoque]
      summary: Consulta o estoque de um produto
      operationId: getStock
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      responses:
        '200':
          description: Produto e seu estoque
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Produto inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Estoque]
      summary: Define a quantidade em estoque de um produto
      operationId: updateStock
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
      responses:
        '200':
          description: Estoque atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Quantidade inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Produto inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stocks/{productId}/increase:
    patch:
      tags: [Estoque]
      summary: Adiciona unidades ao estoque
      operationId: increaseStock
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustStockRequest'
      responses:
        '200':
          description: Estoque incrementado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Quantidade inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Produto inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    Forbidden:
      description: Header x-api-key ausente ou diferente de X_API_KEY
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, description, price, stock]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
          minimum: 0
    CreateProductRequest:
      type: object
      required: [name, slug, price]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
...
    UpdateStockRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 0
    AdjustStockRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
    Error:
      type: object
      required: [error, message, statusCode]
      properties:
        error:
          type: string
          example: ProductNotFoundError
        message:
          type: string
        statusCode:
          type: integer